+++ source/editor/gui/tab/scene-editor/controls/EditorFreeControls.js
@@ -112,7 +112,7 @@ EditorFreeControls.prototype.update = function(mouse, keyboard)
 	// Look camera
 	if (mouse.buttonPressed(Mouse.LEFT))
 	{
-		this.orientation.y -= Editor.settings.editor.mouseLookSensitivity * (Editor.settings.editor.invertNavigation ? mouse.delta.y : -mouse.delta.y);
+		this.orientation.y -= Editor.settings.editor.mouseLookSensitivity * (Editor.settings.editor.invertNavigation ? -mouse.delta.y : mouse.delta.y);
 		this.orientation.x -= Editor.settings.editor.mouseLookSensitivity * mouse.delta.x;
 
 		// Limit Vertical Rotation to 90 degrees